+++ src/views/user/CheckPageLayout.vue
@@ -1,13 +1,18 @@
 <template>
+  <myLoading :active="isCartLoading"></myLoading>
   <FlowChart :outProgress="0"/>
   <div class="container my-4">
     <div class="my-5 row justify-content-center">
       <div class="col-lg-6">
         <h4 class="text-center">購物車資訊</h4>
         <div class="mt-3 text-end">
-          <button type="button" class="btn btn-outline-success" @click="clearCart()">清空購物車</button>
+          <button type="button" class="btn btn-outline-success" @click="clearCart()" :disabled="addBtnState || !cart.total">清空購物車</button>
         </div>
         <hr />
+        <div class="text-center my-5" v-if="!cart.total">
+          <p class="fs-4 fw-bold">你的購物車現在沒東西喔!</p>
+          <router-link to="/products" class="btn btn-lg btn-success">選購商品</router-link>
+        </div>
         <ProductOrderCard v-for="item in cart.carts" :key="item.id" :cart="item"/>
         <div class="row">
           <div class="col-md-6 my-3">
@@ -21,6 +26,7 @@
                 v-model="couponCode"
               />
               <button
+                :disabled="addBtnState || !cart.total"
                 class="btn btn-outline-success"
                 type="button"
                 id="button-addon2"
@@ -39,7 +45,8 @@
       <div class="col-lg-6">
         <h4 class="text-center">訂購人資訊</h4>
         <hr />
-        <v-form @submit="goToPayment()" class="bg-light p-3 rounded" ref="form" v-slot="{ errors }">
+        <v-form @submit="goToPayment()"
+        class="bg-light p-3 rounded" ref="form" v-slot="{ errors }">
           <div class="mb-3">
             <label for="email" class="form-label">Email</label>
             <v-field
@@ -114,7 +121,7 @@
             ></textarea>
           </div>
           <div class="text-end">
-            <button type="submit" class="btn btn-success">
+            <button type="submit" class="btn btn-success" :disabled="addBtnState || !cart.total">
               送出訂單
             </button>
           </div>
@@ -124,32 +131,68 @@
   </div>
 </template>
 <script>
-import { ref } from 'vue'
+import { ref, onMounted } from 'vue'
 import { useRouter } from 'vue-router'
 import FlowChart from '@/components/FlowChart.vue'
 import ProductOrderCard from '@/components/ProductOrderCard.vue'
 import { useCartStore } from '@/stores/cart.js'
 import { storeToRefs } from 'pinia'
 import api from '@/api/axios.js'
+import { notify } from '@/api/toast.js'
+import Swal from 'sweetalert2'
 
 export default {
   components: { FlowChart, ProductOrderCard },
   setup (props) {
     const cartStore = useCartStore()
-    const { cart } = storeToRefs(cartStore)
-    const { getCart, deleteAllCart, useCoupon } = cartStore
+    const { cart, addBtnState, isCartLoading } = storeToRefs(cartStore)
+    const { getCart, deleteAllCart, useCoupon, changeAddStatus } = cartStore
     const useMyCoupon = async () => {
+      changeAddStatus(true)
       const coupon = {
         data: {
           code: couponCode.value
         }
       }
       await useCoupon(coupon)
-      await getCart()
+        .then((res) => {
+          notify(true, res.data.message)
+          getCart()
+        })
+        .catch((err) => {
+          notify(false, err.response.data.message)
+        })
+      changeAddStatus(false)
     }
     const clearCart = async () => {
-      await deleteAllCart()
-      await getCart()
+      Swal.fire({
+        icon: 'warning', // error\warning\info\question
+        title: '確定刪除整個購物車？',
+        text: '刪除後的資料無法恢復',
+        showCancelButton: true,
+        confirmButtonColor: 'red',
+        cancelButtonColor: 'gray',
+        confirmButtonText: '確定',
+        confirmButton: false,
+        cancelButtonText: '取消'
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          changeAddStatus(true)
+          console.log('確認')
+          await deleteAllCart()
+            .then((res) => {
+              notify(true, `全部${res.data.message}`)
+            })
+            .catch((err) => {
+              console.log(err.response.data.message)
+              notify(false, `${err.response.data.message}`)
+            })
+          await getCart()
+          changeAddStatus(false)
+        } else if (result.isDenied) {
+          console.log('取消')
+        }
+      })
     }
     const router = useRouter()
     const couponCode = ref(null)
@@ -180,8 +223,14 @@ export default {
       const phoneNumber = /^(09)[0-9]{8}$/
       return phoneNumber.test(value) ? true : '需要正確的電話號碼'
     }
+    onMounted(() => {
+      changeAddStatus(false)
+    })
     return {
       cart,
+      isCartLoading,
+      addBtnState,
+      changeAddStatus,
       cartStore,
       message,
       couponCode,
